requirements.txt
lxml==5.4.0
pymongo==4.13.0
pyvirtualdisplay==3.0
python3-discogs-client==2.8
python-dotenv==1.1.0
requests==2.32.3
regex==2024.11.6
selenium==4.25.0
webdriver-manager==4.0.2
